requirements.txt
cityscapesscripts==2.2.2
DCNv3.egg==info
detail==0.2.2
einops==0.7.0
fvcore==0.1.5.post20221221
matplotlib==3.7.1
mmcv==2.0.0
mmdet==3.0.0
mmengine==0.7.3
mmpretrain.egg==info
model_archiver==1.0.3
networkx==3.1
nibabel==5.2.0
numpy==1.23.2
numpy==1.24.3
onnx==1.14.0
onnxruntime_gpu==1.8.1
opencv_python==4.7.0.72
oss2==2.18.4
packaging==23.2
Pillow==10.0.0
Pillow==10.2.0
prettytable==3.10.0
ptflops==0.7.1.2
pycocotools==2.0.6
pytest==8.0.1
pytorch_sphinx_theme==0.0.19
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
setuptools==68.0.0
timm==0.9.7
torch==2.0.1+cu117
torchvision==0.15.2+cu117
torchviz==0.0.2
ts==0.5.1
~orch==2.0.1
